Move `Rating` to patterns
