+++ src/components/patterns/Rating/Rating.md
