+++ src/components/patterns/Rating/index.js
